remove GetPtrStorageScope and WithStorageScope from public header
