+++ include/tvm/tir/buffer.h
@@ -199,13 +199,7 @@ class Buffer : public ObjectRef {
 TVM_DLL Buffer decl_buffer(Array<PrimExpr> shape, DataType dtype = DataType::Float(32),
                            String name = "buffer", String storage_scope = "", Span span = Span());
 
-/*!
- * \brief Return the storage scope associated with a buffer variable.
- * \param buffer_var The input buffer variable.
- * \return A string representing the storage scope of this buffer variable.
- */
-TVM_DLL String GetPtrStorageScope(Var buffer_var);
-TVM_DLL Var WithStorageScope(Var buffer_var, String storage_scope);
+
 
 /*!
  * \brief Base node for data producers.
