+++ src/tir/transforms/lower_thread_allreduce.cc
@@ -39,6 +39,13 @@ namespace tir {
 
 namespace {
 
+Var WithStorageScope(Var buffer_var, String storage_scope) {
+  auto* ptr_type = buffer_var->type_annotation.as<PointerTypeNode>();
+  ICHECK(ptr_type) << "The provided variable is not of pointer type";
+  return Var(buffer_var->name_hint, PointerType(ptr_type->element_type, storage_scope),
+             buffer_var->span);
+}
+
 class RemapStorageScope final : public StmtExprMutator {
  public:
   explicit RemapStorageScope(const std::unordered_map<const VarNode*, Var>& new_var_remap)
