+++ src/tir/transforms/lower_warp_memory.cc
@@ -40,11 +40,19 @@
 
 #include "../../arith/pattern_match.h"
 #include "../../runtime/thread_storage_scope.h"
+#include "ir_utils.h"
 
 namespace tvm {
 namespace tir {
 
 namespace {
+
+Var WithStorageScope(Var buffer_var, String storage_scope) {
+  auto* ptr_type = buffer_var->type_annotation.as<PointerTypeNode>();
+  ICHECK(ptr_type) << "The provided variable is not of pointer type";
+  return Var(buffer_var->name_hint, PointerType(ptr_type->element_type, storage_scope),
+             buffer_var->span);
+}
 
 class RemapStorageScope final : public StmtExprMutator {
  public:
